requirements.txt
streamlit==1.35.0
sentence-transformers>=2.7.0
faiss-cpu>=1.9.0
transformers>=4.40.0
torch>=2.3.0
groq>=0.8.0
beautifulsoup4==4.12.3
tiktoken>=0.7.0
rank-bm25>=0.2.2
yfinance>=0.2.40
pandas>=2.2.2
numpy>=1.26.4
plotly>=5.22.0
requests>=2.32.3
python-dotenv>=1.0.1
